+++ src/Recipe/SelectedRecipes.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 
 
-export const SelectedRecipes = ({ recipes, selectedRecipes }) => {
+const SelectedRecipes = ({ recipes, selectedRecipes }) => {
     console.log(selectedRecipes + "d");
 
     // 選択されたレシピの情報を取得
@@ -12,7 +12,7 @@ export const SelectedRecipes = ({ recipes, selectedRecipes }) => {
 
     return (
         <div>
-            <h2>今日の献立</h2>
+            <h2>選択されたレシピ一覧</h2>
             {recipes.length === 0 ? (
                 <p>レシピはありません</p>
             ) : (
@@ -37,3 +37,4 @@ export const SelectedRecipes = ({ recipes, selectedRecipes }) => {
     );
 };
 
+export default SelectedRecipes;
