+++ src/index.js
@@ -1,10 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { RecipePage } from './Recipe';
+import './index.css';
+import App from './App';
+import RecipePage from './Recipe/RecipePage';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
+    {/* <App /> */}
     <RecipePage />
   </React.StrictMode>
 );
